skills/simmersmith-shopping/pyproject.toml
[project]
name = "simmersmith-shopping"
version = "0.1.0"
description = "Local skill that drives Aldi/Walmart/Sam's Club/Instacart carts from a SimmerSmith grocery list."
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.45",
    "httpx>=0.27",
    "pydantic>=2.7",
    "rich>=13.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# PyXA is the modern macOS-only path for reading Reminders. Optional
# so the package can install on non-mac CI; the runtime falls back
# to `osascript` when PyXA is missing or the host isn't macOS.
pyxa = ["pyxa>=0.3.0"]

[project.scripts]
simmersmith-shopping = "simmersmith_shopping.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
